requirements.txt
flask>=3.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
